Add Auth in API

+++ Architectures/Security/TAuth/TAuth.ResourceAPI/Controllers/ResourcesController.cs
@@ -1,9 +1,11 @@
-﻿using Microsoft.AspNetCore.Authorization;
+﻿using IdentityModel;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using TAuth.ResourceAPI.Auth.Policies;
 using TAuth.ResourceAPI.Entities;
 using TAuth.ResourceAPI.Models.Resource;
 
@@ -16,10 +18,13 @@ namespace TAuth.ResourceAPI.Controllers
     public class ResourcesController : ControllerBase
     {
         private readonly ResourceContext _context;
+        private readonly IAuthorizationService _authorizationService;
 
-        public ResourcesController(ResourceContext context)
+        public ResourcesController(ResourceContext context,
+            IAuthorizationService authorizationService)
         {
             _context = context;
+            _authorizationService = authorizationService;
         }
 
         [HttpGet]
@@ -37,11 +42,13 @@ namespace TAuth.ResourceAPI.Controllers
         [HttpGet("{id}")]
         public async Task<IActionResult> Get(int id)
         {
-            var item = await _context.Resources.Select(o => new ResourceDetailModel
-            {
-                Id = o.Id,
-                Name = o.Name
-            }).SingleOrDefaultAsync(o => o.Id == id);
+            var userId = int.Parse(User.FindFirst(JwtClaimTypes.Subject).Value);
+            var item = await _context.Resources.Where(o => o.Id == id && o.OwnerId == userId)
+                .Select(o => new ResourceDetailModel
+                {
+                    Id = o.Id,
+                    Name = o.Name
+                }).SingleOrDefaultAsync();
 
             if (item == null)
                 return NotFound();
@@ -50,6 +57,7 @@ namespace TAuth.ResourceAPI.Controllers
         }
 
         [HttpPost]
+        [Authorize("IsLucky")]
         public async Task<int> Post([FromBody] CreateResourceModel model)
         {
             var entity = new ResourceEntity
@@ -64,16 +72,27 @@ namespace TAuth.ResourceAPI.Controllers
         }
 
         [HttpDelete("{id}")]
-        [Authorize(Roles = "Administrator")]
+        //[Authorize(Roles = "Administrator")]
         public async Task<IActionResult> Delete(int id)
         {
-            var item = await _context.Resources.Select(o => new ResourceEntity
-            {
-                Id = o.Id
-            }).SingleOrDefaultAsync(o => o.Id == id);
+            var item = await _context.Resources.IgnoreQueryFilters()
+                .Where(o => o.Id == id).Select(o => new ResourceEntity
+                {
+                    Id = o.Id,
+                    Name = o.Name,
+                    OwnerId = o.OwnerId
+                }).SingleOrDefaultAsync();
 
             if (item == null)
                 return NotFound();
+
+            var authResult = await _authorizationService.AuthorizeAsync(User, new ResourceAuthorizationModel
+            {
+                Name = item.Name,
+                OwnerId = item.OwnerId
+            }, "CanDeleteResource");
+
+            if (!authResult.Succeeded) return Forbid();
 
             _context.Resources.Remove(item);
             await _context.SaveChangesAsync();
